fixed warning
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@6845 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ mrt/xml.h
@@ -45,7 +45,7 @@ public:
 		}
 		inline float get(const std::string &name, float defv) const {
 			const_iterator i;
-			return (i = find(name)) != end()? atof(i->second.c_str()): defv;
+			return (i = find(name)) != end()? (float)atof(i->second.c_str()): defv;
 		}
 	};
